+++ src/main/scala/pl/polsl/fastq/mode/Mode.scala
@@ -1,4 +1,4 @@
-package pl.polsl.fastq.trimmer
+package pl.polsl.fastq.mode
 
 trait Mode {
   def run(argsMap: Map[String, Any]): Unit
