+++ src/main/scala/pl/polsl/fastq/mode/PairedEndMode.scala
@@ -1,4 +1,4 @@
-package pl.polsl.fastq.trimmer
+package pl.polsl.fastq.mode
 
 class PairedEndMode extends Mode {
   override def run(argsMap: Map[String, Any]): Unit = ???
